+++ lib/src/wave.dart
@@ -5,11 +5,13 @@ import 'package:flutter/material.dart';
 class Wave extends StatefulWidget {
   final double value;
   final Color color;
+  final Axis direction;
 
   const Wave({
     Key key,
     @required this.value,
     @required this.color,
+    @required this.direction,
   }) : super(key: key);
 
   @override
@@ -50,6 +52,7 @@ class _WaveState extends State<Wave> with SingleTickerProviderStateMixin {
             clipper: _WaveClipper(
               animationValue: _animationController.value,
               value: widget.value,
+              direction: widget.direction,
             ),
           ),
     );
@@ -59,20 +62,34 @@ class _WaveState extends State<Wave> with SingleTickerProviderStateMixin {
 class _WaveClipper extends CustomClipper<Path> {
   final double animationValue;
   final double value;
+  final Axis direction;
 
-  _WaveClipper({this.animationValue, this.value});
+  _WaveClipper({
+    @required this.animationValue,
+    @required this.value,
+    @required this.direction,
+  });
 
   @override
   Path getClip(Size size) {
+    if (direction == Axis.horizontal) {
+      Path path = Path()
+        ..addPolygon(_generateVerticalWavePath(size), false)
+        ..lineTo(0.0, size.height)
+        ..lineTo(0.0, 0.0)
+        ..close();
+      return path;
+    }
+
     Path path = Path()
-      ..addPolygon(_generateWavePath(size), false)
+      ..addPolygon(_generateHorizontalWavePath(size), false)
       ..lineTo(size.width, size.height)
       ..lineTo(0.0, size.height)
       ..close();
     return path;
   }
 
-  List<Offset> _generateWavePath(Size size) {
+  List<Offset> _generateHorizontalWavePath(Size size) {
     final waveList = <Offset>[];
     for (int i = -2; i <= size.width.toInt() + 2; i++) {
       final waveHeight = (size.height / 20);
@@ -80,6 +97,18 @@ class _WaveClipper extends CustomClipper<Path> {
               waveHeight +
           (size.height - (size.height * value));
       waveList.add(Offset(i.toDouble(), dy));
+    }
+    return waveList;
+  }
+
+  List<Offset> _generateVerticalWavePath(Size size) {
+    final waveList = <Offset>[];
+    for (int i = -2; i <= size.height.toInt() + 2; i++) {
+      final waveHeight = (size.width / 20);
+      final dx = math.sin((animationValue * 360 - i) % 360 * (math.pi / 180)) *
+              waveHeight +
+          (size.width - (size.width * value));
+      waveList.add(Offset(dx, i.toDouble()));
     }
     return waveList;
   }